fixing bugs

+++ P2_840128088/src/dataManagementClasses/DoubleDataWriter.java
@@ -43,6 +43,6 @@ public class DoubleDataWriter implements DataWriter {
 
 	@Override
 	public String toString(Object value) {
-		return String.format(DataUtils.FLOATFORMAT, (Long) value); 
+		return String.format(DataUtils.FLOATFORMAT, (Double) value); 
 	}
 }
